map texture

+++ Math.cpp
@@ -25,15 +25,15 @@ Vector2f Math::normalizeVector(Vector2f vector)
 //}
 bool Math::checkRectCollision(FloatRect rect1, FloatRect rect2)
 {	
-	int rect1Left = rect1.position.x;
-	int rect1Right = rect1Left + rect1.size.x;
-	int rect2Left  = rect2.position.x;
-	int rect2Right = rect2Left + rect2.size.x;
+	int rect1Left = (int)rect1.position.x;
+	int rect1Right = (int)rect1Left + rect1.size.x;
+	int rect2Left  = (int)rect2.position.x;
+	int rect2Right = (int)rect2Left + rect2.size.x;
 
-	int rect1Top = rect1.position.y;
-	int rect1Bottom = rect1Top + rect1.size.y;
-	int rect2Top = rect2.position.y;
-	int rect2Bottom = rect2Top + rect2.size.y;
+	int rect1Top = (int)rect1.position.y;
+	int rect1Bottom = (int)rect1Top + rect1.size.y;
+	int rect2Top = (int)rect2.position.y;
+	int rect2Bottom = (int)rect2Top + rect2.size.y;
 
 	if (rect2Left < rect1Right &&
 		rect2Right > rect1Left &&
